requirements.txt
APScheduler==3.11.0
fastapi==0.115.12
grpcio==1.71.0
minio==7.2.15
numpy==1.26.3
opencv_python==4.8.1.78
Pillow==10.2.0
pika==1.3.2
pydantic==2.11.4
pydantic_settings==2.9.1
pytest==8.3.4
python-dotenv==1.1.0
python_jose==3.3.0
Requests==2.32.3
httpx[http2]==0.28.1
SQLAlchemy==2.0.25
tritonclient[all]==2.41.0
uvicorn==0.34.2
google==3.0.0
protobuf==5.29.4
pymysql==1.1.1
python-multipart==0.0.19
scipy==1.16.0
scikit-image==0.25.2
lap==0.5.12
filterpy==1.4.5
langchain==0.3.26
langchain-openai==0.3.27
# langchain-anthropic==0.2.5
# langchain-google-genai==2.0.5
# langchain-community==0.3.10
langchain-core==0.3.66
redis==5.2.1
cryptography
# JWT和认证相关
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
# PyAV - 高稳定性RTSP推流
av>=15.0.0
psutil==7.0.0